packages/erasure-executor/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[project]
name = "erasure-executor"
version = "0.1.0"
requires-python = ">=3.11,<3.13"
dependencies = [
  "fastapi==0.129.0",
  "uvicorn==0.41.0",
  "pydantic==2.12.5",
  "PyYAML==6.0.3",
  "SQLAlchemy==2.0.46",
  "psycopg[binary]==3.2.3",
  "alembic==1.18.4",
  "prometheus-client==0.24.1",
  "httpx==0.28.1",
  "beautifulsoup4==4.14.3",
  "lxml==6.0.2",
  "jsonschema==4.25.1",
  "cryptography>=44.0.0",
  "playwright>=1.50.0",
]

[project.optional-dependencies]
test = ["pytest==8.4.0", "pytest-asyncio==0.26.0"]
lint = ["ruff==0.9.10", "mypy==1.15.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
